+++ src/conversion.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { days, hours, minutes, seconds } from './conversion.js';
+import { days, hours, minutes, seconds, weeks } from './conversion.js';
 import { UnitConverter } from './utils/createUnitConverter.js';
 
 const testCases: [unit: string, fn: UnitConverter, milliseconds: number][] = [
@@ -7,6 +7,7 @@ const testCases: [unit: string, fn: UnitConverter, milliseconds: number][] = [
 	['minutes', minutes, 60_000],
 	['hours', hours, 3_600_000],
 	['days', days, 86_400_000],
+	['weeks', weeks, 604_800_000],
 ];
 
 const testUnitAmounts = [1, 2, 3, 5, 8];