+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `weeks` unit converter
+- `WEEK` and `DAYS_IN_WEEK` constants
+
 ### Fixes
 
 - README typos